refactor: merge all contexts into one
