+++ example/src/index.tsx
@@ -1,19 +1,9 @@
-import {
-  DMJobProvider,
-  DMSSProvider,
-  UiPluginProvider,
-} from '@development-framework/dm-core'
 import React from 'react'
-import { ToastContainer } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.min.css'
 
 import ReactDOM from 'react-dom/client'
 import { AuthProvider } from 'react-oauth2-code-pkce'
 import App from './App'
-import plugins from './plugins'
-
-const fullCurrentURL = () =>
-  `${window.location.pathname}${window.location.search}${window.location.hash}`
 
 const authEnabled = import.meta.env.VITE_AUTH_ENABLED === '1'
 const authConfig = {
@@ -31,34 +21,16 @@ const authConfig = {
   }/oauth2/logout`,
 }
 
-const Content = () => {
-  const enableBlueprintCache =
-    import.meta.env.VITE_BLUEPRINT_CACHE_ENABLED === '1' || true
-  return (
-    <DMSSProvider
-      dmssBasePath={import.meta.env.VITE_DMSS_URL}
-      enableBlueprintCache={enableBlueprintCache}
-    >
-      <DMJobProvider dmJobPath={import.meta.env.VITE_DM_JOB_URL}>
-        <UiPluginProvider pluginsToLoad={plugins}>
-          <App />
-        </UiPluginProvider>
-        <ToastContainer />
-      </DMJobProvider>
-    </DMSSProvider>
-  )
-}
-
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement)
 
 root.render(
   <React.StrictMode>
     {authEnabled ? (
       <AuthProvider authConfig={authConfig}>
-        <Content />
+        <App />
       </AuthProvider>
     ) : (
-      <Content />
+      <App />
     )}
   </React.StrictMode>
 )
