+++ packages/dm-core/src/index.tsx
@@ -6,12 +6,7 @@ export { AuthContext, AuthProvider } from 'react-oauth2-code-pkce'
 export * from './Enums'
 export * from './components'
 export { colors } from './colors'
-export * from './context/ApplicationContext'
-export * from './context/DMSSContext'
-export * from './context/DMJobContext'
-export * from './context/FileSystemTreeContext'
-export * from './context/UiPluginContext'
-export * from './context/RoleContext'
+export * from './ApplicationContext'
 export * from './domain/Tree'
 export * from './services'
 export * from './services/api/configs/gen-job/models'
